with new run_suite.py
